requirements.txt
pandas
thefuzz[speedup]
python-dateutil
pydantic
langchain-google-genai